chore: 🤖 Product immutable 수정
- setter 메소드를 제거하여 product immutable

✅ Closes: #40

+++ src/main/java/com/giftclub/mapper/ProductMapper.java
@@ -10,5 +10,5 @@ public interface ProductMapper {
 
     public Product selectProductById(Long id);
 
-    public void decreaseProduct(Long id, int stock);
+    public void decreaseProduct(Long id, Long stock);
 }
